очищен код, исправлен double-> int m_count изменен Mean::update

+++ homework_5/Src/Mean.cpp
@@ -1,18 +1,15 @@
 #include"Mean.h"
 
-void Mean::update(double next)
-{
-		m_sum += next;
-		m_mean = m_sum / m_count;
-		m_count++;
-	}
+void Mean::update(double next){
+	m_sum += next;
+	m_mean = m_sum/static_cast<double>(m_count);
+	m_count++;
+}
 
-double Mean::eval() const
-{
+double Mean::eval() const{
 	return m_mean;
 }
 
-const char* Mean::name() const
-{
+const char* Mean::name() const{
 	return "Mean";
 }
